Fixed network writing.
git-svn-id: https://svn.code.sf.net/p/stemkit/code@1497 faf13d3b-af0d-0410-97ba-fcc57a00092e

+++ lib/sk/net/DataOutputStream.cc
@@ -65,12 +65,12 @@ writeLong(long long value)
   if(_bigEndian == false) {
     x = ((((uint64_t)htonl(x)) << 32) + htonl(x >> 32));
   }
-  sk::io::DataOutputStream::writeInt(x);
+  sk::io::DataOutputStream::writeLong(x);
 }
 
 void 
 sk::net::DataOutputStream::
 writeShort(short value)
 {
-  sk::io::DataOutputStream::writeInt(htons(value));
+  sk::io::DataOutputStream::writeShort(htons(value));
 }
